add endpoint type validation

+++ hermes-portal/src/main/java/com/ctrip/hermes/portal/service/ConsumerService.java
@@ -10,6 +10,7 @@ import org.unidal.lookup.annotation.Inject;
 import org.unidal.lookup.annotation.Named;
 
 import com.ctrip.hermes.meta.entity.ConsumerGroup;
+import com.ctrip.hermes.meta.entity.Endpoint;
 import com.ctrip.hermes.meta.entity.Meta;
 import com.ctrip.hermes.meta.entity.Topic;
 import com.ctrip.hermes.metaservice.service.MetaServiceWrapper;
@@ -54,7 +55,7 @@ public class ConsumerService {
 		ConsumerGroup consumerGroup = t.findConsumerGroup(consumer);
 		if (consumerGroup != null) {
 			boolean removed = t.removeConsumerGroup(consumer);
-			if (removed) {
+			if (removed && Endpoint.BROKER.equals(t.getEndpointType())) {
 				m_storageService.delConsumerStorage(t, consumerGroup);
 				m_zookeeperService.deleteConsumerLeaseZkPath(t, consumer);
 			}
@@ -77,9 +78,10 @@ public class ConsumerService {
 		Topic t = meta.getTopics().get(topic);
 		t.addConsumerGroup(consumer);
 
-		m_storageService.addConsumerStorage(t, consumer);
-
-		m_zookeeperService.ensureConsumerLeaseZkPath(t);
+		if (Endpoint.BROKER.equals(t.getEndpointType())) {
+			m_storageService.addConsumerStorage(t, consumer);
+			m_zookeeperService.ensureConsumerLeaseZkPath(t);
+		}
 
 		if (!m_metaService.updateMeta(meta)) {
 			throw new RuntimeException("Update meta failed, please try later");
